+++ sentry-spring-boot-4/src/test/kotlin/com/acme/MainBootClass.kt
@@ -1,0 +1,5 @@
+package com.acme
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+
+@SpringBootApplication open class MainBootClass